refactor: EntityManager 에서 Mapper을 관리하도록 수정
