+++ src/main/java/persistence/sql/entity/manager/EntityManagerImpl.java
@@ -4,10 +4,9 @@ import jdbc.JdbcTemplate;
 import persistence.sql.dml.query.builder.DeleteQueryBuilder;
 import persistence.sql.dml.query.builder.InsertQueryBuilder;
 import persistence.sql.dml.query.builder.SelectQueryBuilder;
-import persistence.repository.RepositoryMapper;
 import persistence.sql.entity.EntityMappingTable;
-import persistence.sql.entity.conditional.Criterion;
 import persistence.sql.entity.conditional.Criteria;
+import persistence.sql.entity.conditional.Criterion;
 import persistence.sql.entity.model.DomainType;
 import persistence.sql.entity.model.Operators;
 
@@ -17,31 +16,33 @@ import java.util.List;
 public class EntityManagerImpl<T, K> implements EntityManger<T, K> {
 
     private final JdbcTemplate jdbcTemplate;
+    private final EntityManagerMapper<T> entityManagerMapper;
 
-    public EntityManagerImpl(final JdbcTemplate jdbcTemplate) {
+    public EntityManagerImpl(final JdbcTemplate jdbcTemplate,
+                             Class<T> clazz) {
         this.jdbcTemplate = jdbcTemplate;
+        this.entityManagerMapper = new EntityManagerMapper<>(clazz);
     }
+
 
     @Override
     public List<T> findAll(final Class<T> clazz) {
         final EntityMappingTable entityMappingTable = EntityMappingTable.from(clazz);
-        final RepositoryMapper<T> repositoryMapper = new RepositoryMapper<>(clazz);
 
         SelectQueryBuilder selectQueryBuilder = SelectQueryBuilder.from(entityMappingTable);
-        return jdbcTemplate.query(selectQueryBuilder.toSql(), repositoryMapper::mapper);
+        return jdbcTemplate.query(selectQueryBuilder.toSql(), entityManagerMapper::mapper);
     }
 
     @Override
     public T find(final Class<T> clazz, final K id) {
         final EntityMappingTable entityMappingTable = EntityMappingTable.from(clazz);
-        final RepositoryMapper<T> repositoryMapper = new RepositoryMapper<>(clazz);
         final DomainType pkDomainType = entityMappingTable.getPkDomainTypes();
 
         Criterion criterion = new Criterion(pkDomainType.getColumnName(), id.toString(), Operators.EQUALS);
         Criteria criteria = new Criteria(Collections.singletonList(criterion));
 
         SelectQueryBuilder selectQueryBuilder = SelectQueryBuilder.of(entityMappingTable, criteria);
-        return jdbcTemplate.queryForObject(selectQueryBuilder.toSql(), repositoryMapper::mapper);
+        return jdbcTemplate.queryForObject(selectQueryBuilder.toSql(), entityManagerMapper::mapper);
     }
 
     @Override
@@ -54,9 +55,8 @@ public class EntityManagerImpl<T, K> implements EntityManger<T, K> {
     public void remove(final T entity) {
         final EntityMappingTable entityMappingTable = EntityMappingTable.from(entity.getClass());
         final DomainType pkDomainType = entityMappingTable.getPkDomainTypes();
-        final EntityManagerMapper<T> entityManagerMapper = new EntityManagerMapper(entity, entity.getClass(), pkDomainType.getName());
 
-        Criterion criterion = new Criterion(pkDomainType.getColumnName(), entityManagerMapper.getFieldValue(), Operators.EQUALS);
+        Criterion criterion = new Criterion(pkDomainType.getColumnName(), entityManagerMapper.getFieldValue(entity, pkDomainType.getColumnName()), Operators.EQUALS);
         Criteria criteria = new Criteria(Collections.singletonList(criterion));
 
         DeleteQueryBuilder deleteQueryBuilder = DeleteQueryBuilder.of(entityMappingTable.getTableName(), criteria);
